move genaral types into sharedd folder
