--- client/src/entities/group/index.ts
@@ -1,1 +1,0 @@
-export type { Group } from "./model/types";
