--- client/src/entities/member/index.ts
@@ -1,1 +1,0 @@
-export type { Member, MemberWithName } from "./model/types";
