--- client/src/entities/message/index.ts
@@ -1,1 +1,0 @@
-export type { Message, MessageWithName, ReadyMessage } from "./model/types";
