+++ client/src/shared/api/message/sendMessage.ts
@@ -1,5 +1,5 @@
 import { axiosInstance, API_URL } from "@/shared/config/axiosInstance";
-import type { ReadyMessage } from "@/entities/message/model/types";
+import type { ReadyMessage } from "@/shared/types/Message";
 
 export const apiSendMessage = async (message: ReadyMessage, room: number) => {
     const res = await axiosInstance.post(`${API_URL}/send-message`, {
